AEAZ-6830 - change salesOrderQty from int to double

+++ pangea-parent/pangea-view/src/main/java/com/jnj/pangea/GDMDemandSalesOrder.java
@@ -369,7 +369,7 @@ public class GDMDemandSalesOrder implements IEventProcessor {
 
 		if (inclusionExclusion != null) {
 			if (Integer.parseInt(scheduleLineItem) != 1
-					|| Integer.parseInt(salesOrderQty) <= 0
+					|| Double.parseDouble(salesOrderQty) <= 0
 					|| StringInner.isStringNotEmpty(localRejReason)
 					|| "C".equals(lineRejStat)) {
 				return false;
